+++ test/e2e-build.sh
@@ -11,5 +11,10 @@ cd ${REPO_ROOT} && docker build -t test/flagger:latest . -f Dockerfile
 echo '>>> Installing Flagger'
 kind load docker-image test/flagger:latest
 kubectl apply -f ${REPO_ROOT}/artifacts/flagger/
+
+if [ -n "$1" ]; then
+  kubectl -n istio-system set env deployment/flagger "MESH_PROVIDER=$1"
+fi
+
 kubectl -n istio-system set image deployment/flagger flagger=test/flagger:latest
 kubectl -n istio-system rollout status deployment/flagger