+++ test/e2e-nginx-tests.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 # This script runs e2e tests for Canary initialization, analysis and promotion
-# Prerequisites: Kubernetes Kind, Helm and Istio
+# Prerequisites: Kubernetes Kind, Helm and NGINX ingress controller
 
 set -o errexit
 
@@ -10,14 +10,14 @@ export KUBECONFIG="$(kind get kubeconfig-path --name="kind")"
 
 echo '>>> Creating test namespace'
 kubectl create namespace test
-kubectl label namespace test istio-injection=enabled
 
-echo '>>> Installing the load tester'
+echo '>>> Installing load tester'
 kubectl -n test apply -f ${REPO_ROOT}/artifacts/loadtester/
 kubectl -n test rollout status deployment/flagger-loadtester
 
 echo '>>> Initialising canary'
 kubectl apply -f ${REPO_ROOT}/test/e2e-workload.yaml
+kubectl apply -f ${REPO_ROOT}/test/e2e-ingress.yaml
 
 cat <<EOF | kubectl apply -f -
 apiVersion: flagger.app/v1alpha3
@@ -30,15 +30,13 @@ spec:
     apiVersion: apps/v1
     kind: Deployment
     name: podinfo
+  ingressRef:
+    apiVersion: extensions/v1beta1
+    kind: Ingress
+    name: podinfo
   progressDeadlineSeconds: 60
   service:
     port: 9898
-    headers:
-      request:
-        add:
-          x-envoy-upstream-rq-timeout-ms: "15000"
-          x-envoy-max-retries: "10"
-          x-envoy-retry-on: "gateway-error,connect-failure,refused-stream"
   canaryAnalysis:
     interval: 15s
     threshold: 15
@@ -48,40 +46,27 @@ spec:
     - name: request-success-rate
       threshold: 99
       interval: 1m
-    - name: request-duration
-      threshold: 500
-      interval: 30s
-    - name: "404s percentage"
-      threshold: 5
+    - name: "latency"
+      threshold: 0.5
       interval: 1m
       query: |
-        100 - sum(
+        histogram_quantile(0.99,
+          sum(
             rate(
-                istio_requests_total{
-                  reporter="destination",
-                  destination_workload_namespace=~"test",
-                  destination_workload=~"podinfo",
-                  response_code!="404"
-                }[1m]
+              http_request_duration_seconds_bucket{
+                kubernetes_namespace="test",
+                kubernetes_pod_name=~"podinfo-[0-9a-zA-Z]+(-[0-9a-zA-Z]+)"
+              }[1m]
             )
+          ) by (le)
         )
-        /
-        sum(
-            rate(
-                istio_requests_total{
-                  reporter="destination",
-                  destination_workload_namespace=~"test",
-                  destination_workload=~"podinfo"
-                }[1m]
-            )
-        ) * 100
     webhooks:
       - name: load-test
         url: http://flagger-loadtester.test/
         timeout: 5s
         metadata:
           type: cmd
-          cmd: "hey -z 10m -q 10 -c 2 http://podinfo.test:9898/"
+          cmd: "hey -z 10m -q 10 -c 2 -host app.example.com http://nginx-ingress-controller.ingress-nginx"
           logCmdOutput: "true"
 EOF
 
@@ -94,7 +79,7 @@ until ${ok}; do
     sleep 5
     count=$(($count + 1))
     if [[ ${count} -eq ${retries} ]]; then
-        kubectl -n istio-system logs deployment/flagger
+        kubectl -n ingress-nginx logs deployment/flagger
         echo "No more retries left"
         exit 1
     fi
@@ -112,12 +97,12 @@ ok=false
 until ${ok}; do
     kubectl -n test describe deployment/podinfo-primary | grep '1.4.1' && ok=true || ok=false
     sleep 10
-    kubectl -n istio-system logs deployment/flagger --tail 1
+    kubectl -n ingress-nginx logs deployment/flagger --tail 1
     count=$(($count + 1))
     if [[ ${count} -eq ${retries} ]]; then
         kubectl -n test describe deployment/podinfo
         kubectl -n test describe deployment/podinfo-primary
-        kubectl -n istio-system logs deployment/flagger
+        kubectl -n ingress-nginx logs deployment/flagger
         echo "No more retries left"
         exit 1
     fi
@@ -140,6 +125,10 @@ spec:
     apiVersion: apps/v1
     kind: Deployment
     name: podinfo
+  ingressRef:
+    apiVersion: extensions/v1beta1
+    kind: Ingress
+    name: podinfo
   progressDeadlineSeconds: 60
   service:
     port: 9898
@@ -149,15 +138,15 @@ spec:
     iterations: 5
     match:
       - headers:
+          x-canary:
+            exact: "insider"
+      - headers:
           cookie:
-            regex: "^(.*?;)?(type=insider)(;.*)?$"
+            exact: "canary"
     metrics:
     - name: request-success-rate
       threshold: 99
       interval: 1m
-    - name: request-duration
-      threshold: 500
-      interval: 30s
     webhooks:
       - name: pre
         type: pre-rollout
@@ -165,7 +154,7 @@ spec:
         timeout: 5s
         metadata:
           type: cmd
-          cmd: "hey -z 10m -q 10 -c 2 -H 'Cookie: type=insider' http://podinfo-canary.test:9898/"
+          cmd: "hey -z 10m -q 10 -c 2 -H 'X-Canary: insider' -host app.example.com http://nginx-ingress-controller.ingress-nginx"
           logCmdOutput: "true"
       - name: post
         type: post-rollout
@@ -173,7 +162,7 @@ spec:
         timeout: 15s
         metadata:
           type: cmd
-          cmd: "curl -s http://podinfo.test:9898/"
+          cmd: "curl -sH 'Host: app.example.com' http://nginx-ingress-controller.ingress-nginx"
           logCmdOutput: "true"
 EOF
 
@@ -187,12 +176,12 @@ ok=false
 until ${ok}; do
     kubectl -n test describe deployment/podinfo-primary | grep '1.4.2' && ok=true || ok=false
     sleep 10
-    kubectl -n istio-system logs deployment/flagger --tail 1
+    kubectl -n ingress-nginx logs deployment/flagger --tail 1
     count=$(($count + 1))
     if [[ ${count} -eq ${retries} ]]; then
         kubectl -n test describe deployment/podinfo
         kubectl -n test describe deployment/podinfo-primary
-        kubectl -n istio-system logs deployment/flagger
+        kubectl -n ingress-nginx logs deployment/flagger
         echo "No more retries left"
         exit 1
     fi
@@ -200,6 +189,6 @@ done
 
 echo '✔ A/B testing promotion test passed'
 
-kubectl -n istio-system logs deployment/flagger
+kubectl -n ingress-nginx logs deployment/flagger
 
 echo '✔ All tests passed'